요금메뉴 insert, update 구분 방법 변경

+++ app/src/main/java/com/youngs/mygym/ui/employee/member/room/MemberEntity.kt
@@ -16,6 +16,6 @@ data class MemberEntity(
     var remark: String
 ){
     @PrimaryKey(autoGenerate = true)
-    var index: Long = 0
+    var index: Int = 0
 }
